services/arbitrage-engine/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
numpy
pydantic
